Refactor 2-axis servo control and setup
This commit refactors the code for controlling the 2-axis servo in the `2-axis_servo.h` file. It introduces separate functions for moving the servo left, right, up, and down. The starting positions for the servos are now set to the center (90 degrees). The code also corrects the usage of the `server.on` function in the `setup_2servo` function in the `esp32-cam.ino` file.